+++ src/app/error.tsx
@@ -70,4 +70,5 @@ const errMap: { [key: string]: string } = {
   '이미 존재하는 상품명입니다.': '이미 존재하는 영화명입니다',
   '유저를 찾을 수 없습니다.': '존재하지 않는 유저입니다',
   '상품을 찾을 수 없습니다.': '상품을 찾을 수 없습니다',
+  '예고편이 존재하지 않습니다': '예고편이 존재하지 않습니다',
 };
