+++ src/types/justwatch/providers.ts
@@ -1,0 +1,4 @@
+export interface ProviderInfo {
+  url: string;
+  logo?: string;
+}
